+++ FluentDocker.Tests/CoreTests/Driver/Docker/DockerCliProductionReadinessTests.cs
@@ -235,7 +235,6 @@ exit 0
 
       public DockerBinary[] Binaries => [_binary];
       public DockerBinary MainDockerClient => _binary;
-      public DockerBinary MainDockerCompose => _binary;
       public DockerBinary MainDockerCli => _binary;
       public DockerBinary Resolve(string binary) => _binary;
       public string ResolveBinaryPath(string dockerCommand) => _binary.FqPath;