+++ FluentDocker.Tests/CoreTests/Driver/Docker29CliParsingTests.cs
@@ -234,7 +234,6 @@ namespace FluentDocker.Tests.CoreTests.Driver
       private readonly DockerBinary _binary = new(directory, "docker", SudoMechanism.None, null!);
       public DockerBinary[] Binaries => [_binary];
       public DockerBinary MainDockerClient => _binary;
-      public DockerBinary MainDockerCompose => _binary;
       public DockerBinary MainDockerCli => _binary;
       public DockerBinary Resolve(string binary) => _binary;
       public string ResolveBinaryPath(string dockerCommand) => _binary.FqPath;