+++ FluentDocker/Drivers/Docker/Cli/Components/DockerCliContainerDriver.Inspection.cs
@@ -29,7 +29,7 @@ namespace FluentDocker.Drivers.Docker.Cli.Components
     {
       try
       {
-        var result = await ExecuteCommandAsync(context, $"inspect {QuotePositionalArgument(containerId, nameof(containerId))}", cancellationToken).ConfigureAwait(false);
+        var result = await ExecuteCommandAsync(context, $"container inspect {QuotePositionalArgument(containerId, nameof(containerId))}", cancellationToken).ConfigureAwait(false);
 
         if (!result.Success)
         {